LayoutTests/http/wpt/html/semantics/scripting-1/the-script-element/module/module-fails-decoding-then-load-as-classic-script-expected.txt

CONSOLE MESSAGE: SyntaxError: Invalid character '\ufffd'
CONSOLE MESSAGE: SyntaxError: Invalid character '\ufffd'

PASS UTF-16 module script with UTF-16BE BOM, then loaded as classic script (sync)
PASS UTF-16 module script with UTF-16LE BOM, then loaded as classic script (async)
